requirements.txt
Babel==2.6.0
Django==2.1.7
django-crispy-forms==1.7.2
django-phone-field==1.7.0
django-phonenumber-field==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.9.2
phonenumbers==8.10.6
Pillow==5.4.1
pytz==2018.9
